change to double

+++ activation.cpp
@@ -2,7 +2,7 @@
 
 #include <cmath>
 
-static float Sigmoid(float value){
+static double Sigmoid(double value){
 
     return 2.0 / (1.0 + std::exp( -2 * value) -1);
 }
